pyproject.toml
[project]
name = "mandelbrot-visualizer"
version = "0.1.0"
description = "Interactive Mandelbrot set visualizer with area-selection zooming"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "dearpygui>=1.11.1",
    "numpy>=1.24.0",
    "numba>=0.58.0",
    "pillow>=10.0.0",
    "loguru>=0.7.3",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
mandelbrot = "mandelbrot_gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
